Outfactor Tabulator column definitions logic to utils module (#229)
