+++ frontend/vre/record/record.list.view.js
@@ -1,33 +1,7 @@
 import Backbone from "backbone";
-import {properties} from "../utils/record-ontology";
-import {getStringLiteral} from "../utils/jsonld.model";
 import {vreChannel} from "../radio";
 import Tabulator from "tabulator";
-import {columnChooseMenu} from "../utils/tabulator-utils";
-
-const columnProperties = {
-    'edpoprec:title': {
-        visible: true,
-        widthGrow: 5,
-        formatter: 'textarea',
-    },
-    'edpoprec:placeOfPublication': {
-        visible: true,
-    },
-    'edpoprec:dating': {
-        visible: true,
-        widthGrow: 0.5,
-    },
-    'edpoprec:publisherOrPrinter': {
-        visible: true,
-    },
-    'edpoprec:contributor': {
-        visible: true,
-    },
-    'edpoprec:activity': {
-        visible: true,
-    },
-};
+import {adjustDefinitions} from "../utils/tabulator-utils";
 
 export var RecordListView = Backbone.View.extend({
     id: "record-list",
@@ -48,23 +22,7 @@ export var RecordListView = Backbone.View.extend({
             height: 650, // set height of table (in CSS or here), this enables the Virtual DOM and improves render speed dramatically (can be any valid css height value)
             data: initialData,
             autoColumns: true,
-            autoColumnsDefinitions: (definitions) => {
-                for (let definition of definitions) {
-                    // All columns invisible by default
-                    definition.visible = false;
-                    const property = properties.get(definition.field);
-                    if (property) {
-                        definition.title = getStringLiteral(property.get("skos:prefLabel"));
-                    }
-                    definition.headerFilter = true;
-                    definition.headerContextMenu = columnChooseMenu;
-                    const hardcodedProperties = columnProperties[definition.field];
-                    if (hardcodedProperties) {
-                        Object.assign(definition, hardcodedProperties);
-                    }
-                }
-                return definitions;
-            },
+            autoColumnsDefinitions: adjustDefinitions,
             layout: "fitColumns",
             rowHeader: {
                 width: 50,
